Refactor computation of 2nd poly of `t(X)`
Implemented a function for QuotientToolkit that:

Gets a set of polynomials, passes them to coeficient form
with EvalDomain = 4*n. Then multiplies them by `z(Xw)` and
returns the ifft of the mul over the domain `4n` again.

Refactored `permutation.rs` to work with it on the quotient poly
computation part.