Use HashMap instead of Vec on Permutation
- Refactored the `Permutation` struct to hold
Maps instead of Vecs to store `Variables`
and it's linked `WireData`.
- Refactored `new_variable()` which now adds the
`Variable` in `variables` linked to it's corresponding
`Scalar` and in `variable_map` linked to an empty
`Vec<WireData>` which will hold all of the `WireData` info
related to this `Variable` value.
- Refactor `valid_variables()` to use parallel
iterators and check directly on the `variables`
mapping if the `Variable` is stored there and so,
valid.
- Refactored `add_variable_to_map()` to get
mutable references of the `WireData` vec that corresponds
to a `Variable` and update it with new `WireData` info.
- Refactored `witness_vars_to_scalars()` &
`compute_sigma_permutations()` according to the new
`Permutation` container types.

Added some comments regarding things to discuss about the impl.

This closes #19