Implement benchmarks for Proof Serde impl
- To see if the performance of the de/serialization
can affect the overall Verification process, these benches
were added using Criterion 0.3.0.

- Takes aproximately the same time to serialize and deserialize
a `Proof` and this time is also similar to the time that takes to
verify it. Which would be bad if the time was big but in my Intel I5
it takes 6ms aprox.

- Added Criterion config to `Cargo.toml` and refactored benches.

This provides the bench results requested by #190