Implement padding on perm poly computation
This commit solves & closes #198

We need to pad the witness values since otherways, the len
of them is lower than the evaluation domain size that it's
used.

To put an example:
We preprocess a circuit of size 21 but we pad it to the next
pow of two = 32.

When we try to build proofs with it later, so using the same
prep_circ, it's domain size = 32 but the length of the witnesses
keeps being the same as the circuit size = 21.

This causes the permutation_poly computation function to try to
access indexes of the witness vectors that do not have any
value inside of them.

The solution passes by padding with `Scalar::zero` these vecs.