Fix assertions in logic constraint gates
As mentioned in #251 there was a problem with the `assert`
statements used in the "front-end" check of the PLONK program
memmory.

This was caused by a bit_num parameter set to less bits than the
actual number bit-size.
Therefore, the assert was checking the final value of the resulting
wire agains the original number provided on each input wire instead
of the value clamped to the correct ammount of bits.

Closes #251