Implement a method for constraint-to-const witnes
We needed before to create a witness and then constrain it to a
constant on every circuit we had.
This was time consuming and makes the code less readable since these
kind of constraints should already be performed by the Composer.

- Added the `add_constant_witness` method to the `StandardComposer`
impl block.

Adds the passed `BlsScalar` to the Constraint System as a
witness value and constraints it to be equal to the value that
was sent.

This makes easier to work with Public Inputs, since they can just be
used as witnesses constrained to a fixed value.

Closes #264