Export ProverKey & VerifierKey
We realized during the implementation of https://github.com/dusk-network/dusk-blindbid/issues/63
that we could not read from bytes or a file a `ProverKey` since
the type wasn't exported on the public API of plonk.

Indeed, the visibility of `ProverKey` & `VerifierKey` structures
was `pub(crate)` and this was a problem.