Add lookup tables and hash tables
The lookup tables of different arity
are added to construct lookup gates
and read values for the plookup composer.

The hash table is added for specific values
in the in circuit hashing.

The preprocessed table is used in the creation
of a lookuptable which is shared between the
prover and verifier.

The witness table is where queries will be placed.