Refactor Public Inputs handling in Composer
Previously we were storing the dense representation
of the public inputs that were used with the `Composer`
instance.

This was constly, since most of the public inputs were zero and
therefore storing them was useless but also we needed to collect
the position information in the Circuit trait in such difficult ways.

With this refactor we currently have:
- Public Inputs are now passed to gate-functions as `Option<BlsScalar>`.
- The `Composer` no longer stores the dense public input vector. Instead
we store two sparse vectors which contain the values and the positions.
- We include a function that allows the `Composer` to construct the
dense PI vector from it's sparse ones.
- We've added ways for the consumer to get access to the positions
vector. This will allow to simplify the public input management on the
`Circuit` trait.