Use simple PublicInputValue struct for PI handling
After discussions with @zer0 we realized there's not a need
to have a really big trait to handle the public inputs and all
we need is a simple structure that implements the conversions from the
different PI types that we can have in PLONK circuits into the format
that PLONK requires which is `&[BlsScalar]`.

- Created `PublicInputValue` which is implemented for `BlsScalar`,
`JubJubScalar` and `JubJubAffine`.
- Forced the `Circuit` trait to use this struct as source of the public
input values needed in the verification step.