Make Circuit::TRIM_SIZE auto impl from const_gen
To allow (although it's not recommended to do) variable-size circuits,
we need a way to express different circuit sizes (TRIM_SIZE) for the
same circuit.

By setting the `Circuit::TRIM_SIZE` inherit the value from the const
generic parameter N, we enable this behaviour leaving an API that is not
bad.