Cut visibility of crate structures & fn's
Currently, we expose a way too much things when indeed we should try
to expose publicly in the API only the necessary stuff.

We have modules as fft or transcript_protocol which shouldn't be pub.
Also there are quite some structures which are pub but never meant to
be used by the lib consumers.

The idea is that with this PR we're able to make more clear which are
the things that the consumer is suposed to call or do.

Closes #438