Make fft mod no_std & no_alloc compatible
When we don't have `alloc` we still need to provide the data structures
that don't require it.

Therefore, that's why we've split the mod a bit more and we have a
couple of extra files. On that way is easier to appli the compiler flags
propperly and on a more clear way.