Add `VerifierData` struct
The Public Input positions and the VerifierKey are two items related
logically but not in terms of code. Therefore, it would be nice to also
link them in the code.

By adding `VerifierData` the verifier can always have linked
the pi_positions and the keys of the same circuit linked
with a struct that is serializable.

Resolved: #466