Updated s-box and tests
S-box test expanded to check all outputs are as
expected for various examples. Included a counter
for the s-box that is needed to do the constraint
checks, which will be added next. Also changed bar
related functions to ensure that all variables are
associated with a BlsScalar that is in Montgomery
form, not reduced