Add `identity_point` as method of composer
The current API encapsulates the mutations of the circuit as methods of
the composer.

However, there is an exception with the creation of identity points.

This PR aims to increase the consistency of the API by making available
a method that will allow the user to create an identity point directly
from the composer.

Resolves #572