Add `Witness` as replacement of `AllocatedScalar`
`AllocatedScalar` may suggest to the users to evaluate runtime scalars
to build their gadgets. This is inconsistent because it might create
different circuit descriptions out of the same code.

`AllocatedScalar` was entirely removed in favor of `Witness`. The latter
is a simple index to the witness value in the composer.

Resolves #588