feat: add `compress` to compiler
This commit introduces `Compiler::compress` and `Compiler::decompress`, two
functions that will create a compressed representation of a circuit that
can be used to generate prover and verifier keys.

The compression strategy takes advantage of the fact that circuit
representations are sparse; meaning, most of the scalars are zeroes. We
also have a higher incidence of `1` and `-1`.

This will result in expressive gains in terms of storage. For instance,
a `2^16` circuit can be compressed into roughly 250Kb.